+++ cardano-submit-api/src/Cardano/TxSubmit/Rest/Web.hs
@@ -1,24 +1,22 @@
-{-# LANGUAGE OverloadedStrings #-}
-
 module Cardano.TxSubmit.Rest.Web
   ( runSettings
   )
 where
 
-import           Cardano.Api.Pretty (textShow)
 
-import           Cardano.BM.Trace (Trace, logInfo)
+import           Cardano.Logging.Trace (traceWith)
+import qualified Cardano.Logging.Types as TraceD
+import           Cardano.TxSubmit.Tracing.TraceSubmitApi (TraceSubmitApi (..))
 
 import           Control.Exception (bracket)
 import           Data.Streaming.Network (bindPortTCP)
-import           Data.Text (Text)
 import           Network.Socket (close, getSocketName, withSocketsDo)
 import           Network.Wai.Handler.Warp (Settings, getHost, getPort, runSettingsSocket)
 
 import           Servant (Application)
 
 -- | Like 'Network.Wai.Handler.Warp.runSettings', except with better logging.
-runSettings :: Trace IO Text -> Settings -> Application -> IO ()
+runSettings :: TraceD.Trace IO TraceSubmitApi -> Settings -> Application -> IO ()
 runSettings trace settings app =
   withSocketsDo $
     bracket
@@ -26,6 +24,6 @@ runSettings trace settings app =
       close
       ( \socket -> do
           addr <- getSocketName socket
-          logInfo trace $ "Web API listening on port " <> textShow addr
+          traceWith trace $ EndpointListeningOnPort addr
           runSettingsSocket settings socket app
       )