+++ cardano-submit-api/src/Cardano/TxSubmit/Util.hs
@@ -2,9 +2,8 @@ module Cardano.TxSubmit.Util
   ( logException
   ) where
 
-import           Cardano.Api (textShow)
-
-import           Cardano.BM.Trace (Trace, logError)
+import           Cardano.Logging (Trace, traceWith)
+import           Cardano.TxSubmit.Tracing.TraceSubmitApi (TraceSubmitApi (..))
 
 import           Prelude
 
@@ -15,12 +14,12 @@ import           Data.Text (Text)
 -- code, the caught exception will not be logged. Therefore wrap all tx submission code that
 -- is called from network with an exception logger so at least the exception will be
 -- logged (instead of silently swallowed) and then rethrown.
-logException :: Trace IO Text -> Text -> IO a -> IO a
+logException :: Trace IO TraceSubmitApi -> Text -> IO a -> IO a
 logException tracer txt action = action `catch` logger
   where
     logger :: SomeException -> IO a
     logger e = do
-      logError tracer $ txt <> textShow e
+      traceWith tracer (EndpointException txt e)
       throwIO e